Диплом 2 часть

+++ ewm-service/src/main/java/ru/practicum/explorewithme/server/dto/event/UpdateEventUserRequest.java
@@ -30,7 +30,7 @@ public class UpdateEventUserRequest {
     private LocationDto location;
     private Boolean paid;
     @PositiveOrZero
-    private Long participantLimit;
+    private Integer participantLimit;
     private Boolean requestModeration;
     @Length(min = 3, max = 120)
     @NotBlank
